Move Tag service to `service`
