+++ src/main/scala/com/hhandoko/realworld/route/TagRoutes.scala
@@ -9,7 +9,7 @@ import org.http4s.dsl.Http4sDsl
 import org.http4s.{EntityEncoder, HttpRoutes}
 
 import com.hhandoko.realworld.core.Tag
-import com.hhandoko.realworld.tag.TagService
+import com.hhandoko.realworld.service.TagService
 
 object TagRoutes {
